apps/resume-agent-langgraph/pyproject.toml
[project]
name = "resume-agent-langgraph"
version = "0.3.0"
description = "LangGraph-based Resume Agent MCP Server - Experimental comparison implementation"
authors = [
    {name = "Claude", email = "noreply@anthropic.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "fastmcp>=2.0.0",
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.0",
    "langchain-text-splitters>=0.3.0",
    "sqlite-vec>=0.1.0",
    "httpx>=0.27.0",
    "sqlmodel>=0.0.16",  # Required for DAL imports from apps/resume-agent
    "pyyaml>=6.0.0",  # Required for resume YAML parsing
]

[project.scripts]
resume-agent = "src.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=resume_agent_langgraph",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
